+++ core/src/main/java/org/elasticsearch/Build.java
@@ -19,14 +19,13 @@
 
 package org.elasticsearch;
 
-import org.elasticsearch.common.io.FastStringReader;
-import org.elasticsearch.common.io.Streams;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.joda.time.DateTimeZone;
 import org.joda.time.format.ISODateTimeFormat;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.util.Properties;
 
 /**
@@ -40,10 +39,9 @@ public class Build {
         String hashShort = "NA";
         String timestamp = "NA";
 
-        try {
-            String properties = Streams.copyToStringFromClasspath("/es-build.properties");
+        try (InputStream is = Build.class.getResourceAsStream("/es-build.properties")){
             Properties props = new Properties();
-            props.load(new FastStringReader(properties));
+            props.load(is);
             hash = props.getProperty("hash", hash);
             if (!hash.equals("NA")) {
                 hashShort = hash.substring(0, 7);
