+++ core/src/test/java/org/elasticsearch/action/fieldstats/FieldStatsRequestTest.java
@@ -20,8 +20,8 @@
 package org.elasticsearch.action.fieldstats;
 
 import org.elasticsearch.common.bytes.BytesArray;
-import org.elasticsearch.common.io.Streams;
 import org.elasticsearch.test.ESTestCase;
+import org.elasticsearch.test.StreamsUtils;
 
 import static org.elasticsearch.action.fieldstats.IndexConstraint.Comparison.*;
 import static org.elasticsearch.action.fieldstats.IndexConstraint.Property.MAX;
@@ -31,7 +31,7 @@ import static org.hamcrest.Matchers.equalTo;
 public class FieldStatsRequestTest extends ESTestCase {
 
     public void testFieldsParsing() throws Exception {
-        byte[] data = Streams.copyToBytesFromClasspath("/org/elasticsearch/action/fieldstats/fieldstats-index-constraints-request.json");
+        byte[] data = StreamsUtils.copyToBytesFromClasspath("/org/elasticsearch/action/fieldstats/fieldstats-index-constraints-request.json");
         FieldStatsRequest request = new FieldStatsRequest();
         request.source(new BytesArray(data));
