+++ core/src/test/java/org/elasticsearch/index/mapper/simple/SimpleMapperTests.java
@@ -30,8 +30,8 @@ import org.elasticsearch.index.IndexService;
 import org.elasticsearch.test.ESSingleNodeTestCase;
 import org.junit.Test;
 
-import static org.elasticsearch.common.io.Streams.copyToBytesFromClasspath;
-import static org.elasticsearch.common.io.Streams.copyToStringFromClasspath;
+import static org.elasticsearch.test.StreamsUtils.copyToBytesFromClasspath;
+import static org.elasticsearch.test.StreamsUtils.copyToStringFromClasspath;
 import static org.elasticsearch.index.mapper.MapperBuilders.*;
 import static org.hamcrest.Matchers.equalTo;
