+++ core/src/test/java/org/elasticsearch/index/mapper/update/UpdateMappingTests.java
@@ -36,7 +36,7 @@ import org.junit.Test;
 import java.io.IOException;
 import java.util.LinkedHashMap;
 
-import static org.elasticsearch.common.io.Streams.copyToStringFromClasspath;
+import static org.elasticsearch.test.StreamsUtils.copyToStringFromClasspath;
 import static org.hamcrest.CoreMatchers.equalTo;
